requirements.txt
fastapi 
uvicorn
pydantic
openai
firebase-admin
python-dotenv
google-generativeai
tailwind
